# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cafundo API",
    "version": "3.0.0",
    "description": "API de ordens de serviço, clientes e caixa (Next.js + Prisma + PostgreSQL). Importe este arquivo no Postman via Import > File. Troque o `servers[0].url` (ou a variável `baseUrl` gerada) pelo host desejado. Exemplos usam a massa de teste de usinagem (clientes como \"Metalúrgica Ferrão Ltda\")."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Desenvolvimento local" }
  ],
...
  "paths": {
...
    "/api/ordemServico/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "example": "83a0fa76-0233-405c-9713-4b7d98153557" }
      ],
      "get": {
        "tags": ["Ordem de Serviço"],
        "summary": "Busca ordem por id",
        "responses": {
          "200": { "description": "Ordem encontrada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrdemServico" } } } },
          "404": { "description": "Não encontrada", "content": { "application/json": { "example": { "error": "Ordem de Serviço não encontrada" } } } }
        }
      },
      "put": {
        "tags": ["Ordem de Serviço"],
        "summary": "Atualiza ordem por id",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrdemServicoInput" },
              "example": {
                "cliente": "Usinagem Precisa Tonelli",
                "preco": 980.0,
                "descricao": "Fresamento de flange em ferro fundido GG25",
                "data": "2026-09-05T14:00:00.000Z",
                "maoDeObra": 600.0,
                "material": 380.0,
                "formaDePagamento": "BOLETO_BANCARIO_30"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Ordem atualizada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrdemServico" } } } },
          "500": { "$ref": "#/components/responses/Erro500" }
        }
      },
      "delete": {
        "tags": ["Ordem de Serviço"],
        "summary": "Exclui ordem por id",
        "responses": {
          "200": { "description": "Ordem excluída (retorna o registro)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrdemServico" } } } }
        }
      }
    },
...
    "/api/cliente/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
      ],
      "get": {
        "tags": ["Cliente"],
        "summary": "Busca cliente por id",
        "responses": {
          "200": { "description": "Cliente", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cliente" } } } },
          "404": { "description": "Não encontrado", "content": { "application/json": { "example": { "error": "Cliente não encontrado" } } } }
        }
      },
      "put": {
        "tags": ["Cliente"],
        "summary": "Atualiza cliente por id",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ClienteInput" },
              "example": {
                "nome": "Usinagem Teste Postman Ltda",
                "documento": "99999999000199",
                "telefone": "1133331888",
                "celular": "11999991888",
                "email": "postman@teste.local",
                "endereco": "Rua dos Testes, 100",
                "cidade": "São Paulo",
                "bairro": "Lapa",
                "cep": "05000100"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Cliente atualizado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cliente" } } } }
        }
      },
      "delete": {
        "tags": ["Cliente"],
        "summary": "Exclui cliente por id",
        "responses": {
          "200": { "description": "Cliente excluído (retorna o registro)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cliente" } } } }
        }
      }
    },
...
    "/api/caixa": {
      "get": {
        "tags": ["Caixa"],
        "summary": "Lista transações (404 se vazio)",
        "responses": {
          "200": {
            "description": "Transações",
            "content": {
              "application/json": {
                "example": [
                  { "id": "094682be-1679-4219-a249-a37eb59907a2", "data": "2025-05-04T00:00:00.000Z", "descricao": "Serviço de torneamento - Metalúrgica Ferrão", "categorias": "SERVICOS", "tipo": "ENTRADA", "valor": "1250.75" }
                ]
              }
            }
          },
          "404": { "description": "Sem transações", "content": { "application/json": { "example": { "message": "Nenhuma transação encontrada" } } } }
        }
      },
      "post": {
        "tags": ["Caixa"],
        "summary": "Cria transação",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransacaoInput" },
              "example": {
                "data": "2026-09-06",
                "descricao": "Serviço de fresamento - Fresadora Central do ABC",
                "categorias": "SERVICOS",
                "tipo": "ENTRADA",
                "valor": 1850.0
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Transação criada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransacaoCaixa" } } } },
          "400": { "description": "Campos inválidos", "content": { "application/json": { "example": { "error": "Campos obrigatórios: data, descricao, categorias, tipo, valor" } } } }
        }
      }
    },
    "/api/caixa/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
      ],
      "get": {
        "tags": ["Caixa"],
        "summary": "Busca transação por id",
        "responses": {
          "200": { "description": "Transação", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransacaoCaixa" } } } },
          "404": { "description": "Não encontrada", "content": { "application/json": { "example": { "error": "Transação não encontrada" } } } }
        }
      },
      "put": {
        "tags": ["Caixa"],
        "summary": "Atualiza transação por id",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransacaoInput" },
              "example": {
                "data": "2026-09-06",
                "descricao": "Serviço de fresamento - valor corrigido",
                "categorias": "SERVICOS",
                "tipo": "ENTRADA",
                "valor": 1900.0
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Transação atualizada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransacaoCaixa" } } } }
        }
      },
      "delete": {
        "tags": ["Caixa"],
        "summary": "Exclui transação por id",
        "responses": {
          "200": { "description": "Transação excluída (retorna o registro)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransacaoCaixa" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "OrdemServico": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "cliente": { "type": "string", "example": "Metalúrgica Ferrão Ltda" },
          "preco": { "type": "string", "example": "1250.75", "description": "Decimal serializado como string" },
          "descricao": { "type": "string" },
          "data": { "type": "string", "format": "date-time" },
          "material": { "type": "string", "example": "450.25" },
          "maoDeObra": { "type": "string", "example": "800.50" },
          "formaDePagamento": { "$ref": "#/components/schemas/FormaDePagamento" }
        }
      },
      "OrdemServicoInput": {
        "type": "object",
        "required": ["cliente", "preco", "descricao", "data", "maoDeObra", "material", "formaDePagamento"],
        "properties": {
          "cliente": { "type": "string" },
          "preco": { "type": "number" },
          "descricao": { "type": "string" },
          "data": { "type": "string", "format": "date-time" },
          "maoDeObra": { "type": "number" },
          "material": { "type": "number" },
          "formaDePagamento": { "$ref": "#/components/schemas/FormaDePagamento" }
        }
      },
      "FormaDePagamento": {
        "type": "string",
        "enum": ["CARTAO_DE_CREDITO", "CARTAO_DE_DEBITO", "TRANSFERENCIA_BANCARIA", "BOLETO_BANCARIO", "BOLETO_BANCARIO_10", "BOLETO_BANCARIO_20", "BOLETO_BANCARIO_30", "BOLETO_BANCARIO_60", "BOLETO_BANCARIO_90", "DINHEIRO", "PIX", "OUTROS"]
      },
      "Cliente": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "documento": { "type": "string" },
          "telefone": { "type": "string" },
          "celular": { "type": "string" },
          "email": { "type": "string" },
          "endereco": { "type": "string" },
          "cidade": { "type": "string" },
          "bairro": { "type": "string" },
          "cep": { "type": "string" }
        }
      },
      "ClienteInput": {
        "type": "object",
        "required": ["nome", "documento", "telefone", "celular", "email", "endereco", "cidade", "bairro", "cep"],
        "properties": {
          "nome": { "type": "string" },
          "documento": { "type": "string", "description": "Único" },
          "telefone": { "type": "string" },
          "celular": { "type": "string" },
          "email": { "type": "string", "description": "Único" },
          "endereco": { "type": "string" },
          "cidade": { "type": "string" },
          "bairro": { "type": "string" },
          "cep": { "type": "string" }
        }
      },
      "TransacaoCaixa": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "data": { "type": "string", "format": "date-time" },
          "descricao": { "type": "string" },
          "categorias": { "$ref": "#/components/schemas/Categoria" },
          "tipo": { "$ref": "#/components/schemas/TipoTransacao" },
          "valor": { "type": "string", "example": "1250.75", "description": "Decimal serializado como string" }
        }
      },
      "TransacaoInput": {
        "type": "object",
        "required": ["data", "descricao", "categorias", "tipo", "valor"],
        "properties": {
          "data": { "type": "string", "format": "date", "example": "2026-09-06" },
          "descricao": { "type": "string" },
          "categorias": { "$ref": "#/components/schemas/Categoria" },
          "tipo": { "$ref": "#/components/schemas/TipoTransacao" },
          "valor": { "type": "number" }
        }
      },
      "Categoria": {
        "type": "string",
        "enum": ["VENDAS", "SERVICOS", "INVESTIMENTOS", "REEMBOLSOS", "OUTROS_RECEBIMENTOS", "SALARIO", "ALUGUEL", "MARKETING", "IMPOSTOS", "TRANSPORTE", "FORNECEDORES", "MATERIAIS", "MANUTENCAO", "AGUA", "LUZ", "INTERNET", "OUTRAS_DESPESAS"]
      },
      "TipoTransacao": { "type": "string", "enum": ["ENTRADA", "SAIDA"] },
...
    },
    "responses": {
      "Erro500": {
        "description": "Erro interno",
        "content": { "application/json": { "schema": { "type": "object", "properties": { "error": { "type": "string" } } } } }
      }
    }
  }
}